Add dev notes to Readme.md
Portfolio Project 2 was begun before an initial consultation with
my Mentor, and before receipt of the assesment of Portfolio Project 1

Whilst significant intial development had already been done, it was
decided to start again with a fresh repository build, and try to
incorporate lessons learnt from Mentor and Assesment feedback.

The UXD scope was reworked to break development into stages,and
improve the design. Stage 1 aims to fullfill all Portfolio 2
pass criterea.

Rebuilding the repository afforded the oportunity to try to improve the
use of git commit.