Add outer HTML structure to index.HTML
All content is contained in an outer wrapper set to the width and height
of the viewport.
The Nav section, visable throughout the sight contains the three
elements: Quit, New Game, and Help.
The 2 display sections below contain all of the rest of the site
content. When one section is displayed, the other is hidden.
These sections are styled with columns for responsive layout.
The Readme.md is updated with notes on these features.