Add Nav Bar functionality
Event Listeners were added to the Quit and New Game Nav Bar items.
When Quit is clicked,  main() is called and the welcome screen is
loaded.
When New Game is clicked , newGame() is called, the game screen is
displayed , and when writen, the main game lop will run.

In order to prevent multible image generation when main() was called,
loadImage() was amended to first check for the presence of an image
in its container. If present, the image is removed until the container
was empty. Then the required image is loaded.