Fix map display bug
The map showed a gap between rows. The width of the rows showed
as slightly lrger then the height when inspected in chrome dev tools.
The attribute display:flex was added to the class class 'tile'.
This fixed the issue.
The tile borders were removed.
A default background colour was added to the tile.