Fix Bug Map Over Populating on New Game
The conditional test for pre-existence of map tiles was faulty.
This was corrected. If a tilei mage pre-exisited , it was removed.
The code to create a fresh tile image was encapsulated in
an else statement.

It was assumed that the same error would be present in the code to
display LumberJackie and so was transfered to displayLumberJackie().
However this caused the LumberJackie img to not be created.