Change displayLumberJackie()
Previously displayLumberJackie() iterated through all elMap tiles
to find the current tile.
In this version of displayLumberJackie() the function recieves the
variable currentTileId and displays LumberJackie from this and thus
reduces processing.

displayLumberJackie() is called on loadGame(),
but can be easily called when the current tile is changed (LJ moves),
and can by used for an 'onmouseover' function.