Fix bug: Update Actions Menu List
Next Tile Eventlisteners In Actions Menu
were not  corresponding to new current tile
The issue was related to not being able to remove Eventlisteners.

https://stackoverflow.com/questions/10444077/
javascript-removeeventlistener-not-working was used.
Using an anonymous function call with an event listener made it
impssible to remove an event Listener later.
Changing the method of calling a function as per the stack
overflow solvwd the issue for hoverLumberJackie,
unhoverLumberJackie and the actions menu list.