WIP Bug removeActionEventListeners()
This commit has the most stable code yet: but its still not right.
One harvest tile apears to work ok, but actually calls the sellLogs
function.
A second and subsequent harvesting of a forest tile throw
"currentTile.kind.harvestFoorest is not a function error line 864

For now I need to move on to other issues....