WIP Fix bug hoverLumberJackie()
Bug: Eventlisteners for LumberJackie hove/unhover
fireing multiple times and behaving unpredictably

Partial Solution:
The function hoverLumberJackie() was only acted upon if the eevent
target was the tile image.
 A Html collection of the parent (div)'s children was made.
 This was itterated through. If the iterated element had
 id = "lumber-jackie-hover", this element was removed.
 An image of Lumber-Jackie was then created and put in the tile div.
The function unhoverLumberJackie() m, on filtering the event
tartger to tile image, iterated the same collection and remved the
image of Lumber-Jackie.
Currently, when LJ moves to another tile, a residual LJ image
is left in the exited tile until a unhoverLumberJacket f
unction is called on it.