Fix bug Loading issues after clean up from testing
The function objects logCampActions and harvestForestAction
were being called when the html.index page was refeeshed.
This was because I put them at top scope level main() inorder to
ensure currentTile was always defined. I relocated these two
functions inside the scope of loadGame() and repuplated all the
resulting function parameters eg setTileEventListeners() and
removeTileEventListeners etc.

This resolved the page load issue but did not resolve the issue
that currentTile was at times undefined in harvestForestAction()