Fix bug: harvestForestAction
CurrentTile becomes undefined in harvestForestAction

After 2 or more calls of harvestForestAction currentTile becomes
undefined. This could be caused by something in move(),
and/or a consequence of event listeners not being removed properly.
The chrome DevTools debugger was used to step through the code at
line 342 harvestForestActions until the console error message
was generated.
On first iteration of harvestForestAction, currentTile was define
accurately : in this case as tile object a3. This was true upto and
including the last line 341.
When the function harvestForistAction exited , no return was
specified, but currentTile had been given a new value, and one
was expected later.
Also, because the event listener for Forest Action had not been
removed, and there were possibly by no multiple instances of this
event listener, when harvestForistAction was exited, it was called
again immediately. this time round, currentTile was not defined,
and generated the error on line 339 which required
stockProfit.logsInStock = currentTile.kind.harvestForest(stockProfit)
 to be read.